aulao-1-ia-em-producao/agente-monitoramento/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aulao-monitoring-agent"
version = "0.1.0"
description = "Agente de Monitoramento IA — Aulão IA em Produção"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.0",
    "anthropic>=0.40",
    "openai>=1.30",
    "httpx>=0.27",
    "google-cloud-bigquery>=3.20",
    "google-cloud-logging>=3.10",
    "faiss-cpu>=1.7",
    "numpy>=1.26",
    "structlog>=24.0",
    "rich>=13.0",
    "instructor>=1.7",
    "tenacity>=8.2",
    "click>=8.1",
    "slack-sdk>=3.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = [
    "config",
    "schemas",
    "protocols",
    "providers",
    "pipeline",
    "observability",
    "scripts",
]

[tool.ruff]
target-version = "py313"
line-length = 100
